requirements.txt
Flask==3.0.0
#Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Werkzeug==3.0.1
python-socketio==5.10.0
python-engineio==4.7.1
Pillow==10.1.0
python-dotenv==1.0.0



pip install pyinstaller
pyinstaller --onefile --noconsole run.py
